Modify: chartCotainer API URL 수정, BarChart 파일명 수정
