+++ src/components/charts/player/EGPMChart.js
@@ -3,34 +3,32 @@ import { Bar } from 'react-chartjs-2';
 import styled from 'styled-components';
 import theme from '../../../styles/theme.js';
 
-const ChartBar2 = ({ chartData }) => {
-  let bgc = [
-    theme.red.redB70,
-    theme.blue.blueB70,
-    theme.green.greenB70,
-    theme.orange.orangeB70,
-  ];
-  let data2 = {
-    labels: ['가한 피해량', '받은 피해량', '획득골드', '시야 점수'],
-    datasets: [],
+const EGPMChart = ({ chartData }) => {
+  let data4 = {
+    labels: [],
+    datasets: [
+      {
+        data: [],
+        backgroundColor: [
+          theme.red.redB70,
+          theme.blue.blueB70,
+          theme.green.greenB70,
+          theme.orange.orangeB70,
+        ],
+        borderRadius: 4,
+        barPercentage: 0.8,
+        categoryPercentage: 0.4,
+      },
+    ],
   };
   for (let i = 0; i < chartData.length; i++) {
     let playerName = chartData[i]?.phRole.split('-');
-    data2.datasets.push({
-      label: playerName[3] + ' ' + playerName[4],
-      data: [
-        chartData[i]?.DR,
-        chartData[i]?.DTR,
-        chartData[i]?.GR,
-        chartData[i]?.VSR,
-      ],
-      backgroundColor: bgc[i],
-      borderRadius: 4,
-      barPercentage: 0.8,
-      categoryPercentage: 0.4,
-    });
+    data4.labels.push(playerName[3] + ' ' + playerName[4]);
+    data4.datasets[0].data.push(chartData[i]?.EGPM);
   }
-  const options2 = {
+  const options4 = {
+    suggestedMin: 0,
+    suggestedMax: 100,
     interaction: {
       mode: 'index',
     },
@@ -40,7 +38,7 @@ const ChartBar2 = ({ chartData }) => {
       },
       title: {
         display: true,
-        text: '시즌 평균 동일 포지션 대비 데이터',
+        text: '분당 순수 획득 골드',
         color: theme.white.white100,
         align: 'start',
         padding: '16',
@@ -56,6 +54,7 @@ const ChartBar2 = ({ chartData }) => {
         },
         ticks: {
           color: theme.white.white80,
+          display: false,
         },
       },
       y: {
@@ -67,22 +66,19 @@ const ChartBar2 = ({ chartData }) => {
         ticks: {
           color: theme.white.white80,
         },
-        suggestedMin: 0,
-        suggestedMax: 100,
       },
     },
   };
   return (
     <BarLayout>
-      <Bar data={data2} options={options2} />
+      <Bar data={data4} options={options4} />
     </BarLayout>
   );
 };
 
 const BarLayout = styled.div`
-  width: 548px;
+  width: 200px;
   height: 304px;
-  margin: 0 32px 0 32px;
 `;
 
-export default ChartBar2;
+export default EGPMChart;
